+++ cmd/skywire-cli/commands/visor/root.go
@@ -1,16 +1,12 @@
-package visor
+package clivisor
 
 import (
 	"net"
 	"time"
 
+	"github.com/skycoin/skycoin/src/util/logging"
 	"github.com/spf13/cobra"
 
-	"github.com/skycoin/skywire-utilities/pkg/logging"
-	"github.com/skycoin/skywire/cmd/skywire-cli/commands/visor/vapps"
-	"github.com/skycoin/skywire/cmd/skywire-cli/commands/visor/vroute"
-	"github.com/skycoin/skywire/cmd/skywire-cli/commands/visor/vtp"
-	"github.com/skycoin/skywire/cmd/skywire-cli/commands/visor/vvpn"
 	"github.com/skycoin/skywire/pkg/visor"
 )
 
@@ -19,10 +15,6 @@ var logger = logging.MustGetLogger("skywire-cli")
 var rpcAddr string
 
 func init() {
-	RootCmd.AddCommand(vapps.RootCmd)
-	RootCmd.AddCommand(vroute.RootCmd)
-	RootCmd.AddCommand(vtp.RootCmd)
-	RootCmd.AddCommand(vvpn.RootCmd)
 	RootCmd.PersistentFlags().StringVarP(&rpcAddr, "rpc", "", "localhost:3435", "RPC server address")
 }